+++ src/vws/query.py
@@ -25,9 +25,9 @@ from vws.exceptions.custom_exceptions import (
     RequestEntityTooLargeError,
     ServerError,
 )
-from vws.exceptions.response import Response
 from vws.include_target_data import CloudRecoIncludeTargetData
 from vws.reports import QueryResult, TargetData
+from vws.types import Response
 
 _ImageType = io.BytesIO | BinaryIO
 
@@ -154,6 +154,7 @@ class CloudRecoService:
             status_code=requests_response.status_code,
             headers=dict(requests_response.headers),
             request_body=requests_response.request.body,
+            raw=requests_response.raw,
         )
 
         if response.status_code == HTTPStatus.REQUEST_ENTITY_TOO_LARGE: