+++ CHANGELOG.rst
@@ -4,8 +4,8 @@ Changelog
 Next
 ----
 
-2024.09.03.1
-------------
+* Move ``Response`` from ``vws.exceptions.response`` to ``vws.types``.
+* Add ``raw`` field to ``Response``.
 
 2024.09.03
 ------------